requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
pymysql
python-dotenv